requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
anthropic==0.28.0
httpx==0.27.0
pytest==8.2.1
pytest-asyncio==0.23.7
